clk: meson: axg-audio: migrate to the new parent description method
This file use the string comparison method to describe parent relation
between the clocks, which is not optimized. A recent patch [1] allows
parents to be specified without string names.

Migrate to the new way by using .parent_hws where possible (ie. when
all clocks are local to the controller) and use .parent_data otherwise.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git/commit/?id=fc0c209c147f35ed2648adda09db39fcad89e334&h=master

Signed-off-by: Alexandre Mergnat <amergnat@baylibre.com>