clk: meson: clk-regmap: migrate to new parent description method
This file use the string comparison method to describe parent relation
between the clocks, which is not optimized. A recent patch [1] allows
parents to be specified without string names.

Migrate to the new way in the meson gate macro (located in clk-regmap.h).
Refactor all macro call by adding parent pointer in argument.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git/commit/?id=fc0c209c147f35ed2648adda09db39fcad89e334&h=master

Signed-off-by: Alexandre Mergnat <amergnat@baylibre.com>