clk: meson: meson-eeclk: remove workaround clock
Remove ee-in-xtal's registration which is like the mirror of the
xtal device tree clock. It's was used to provide a xtal global name
but it isn't needed any more thanks to new parent structures in the clock
framework and the parent clocks rework of meson SoC.

Keep the test of the xtal's existence because lot of meson's clocks
registrations need the xtal clock. So in case of xtal DT clock isn't
found, probe function stop.

Signed-off-by: Alexandre Mergnat <amergnat@baylibre.com>